requirements.txt
Flask==3.0.3
matplotlib==3.7.2
yfinance==0.2.36
gunicorn==22.0.0
numpy==1.23.5
finnhub-python==2.4.20
python-dotenv==1.0.0